pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["ycalk <alosev752@gmail.com>"]
packages = [
    {include = "parser/api",from = "bases"},
    {include = "parser/persistence",from = "components"},
    {include = "parser/logging",from = "components"},
    {include = "parser/migrator",from = "bases"},
    {include = "parser/dto",from = "components"},
    {include = "parser/storage",from = "components"},
    {include = "parser/opentele",from = "components"},
    {include = "parser/telegram",from = "components"},
    {include = "parser/scheduler",from = "components"},
    {include = "parser/message_broker",from = "components"},
    {include = "parser/worker",from = "bases"},
    {include = "parser/scheduler_loop",from = "bases"},
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = {version = "^0.128.0", extras = ["standard"]}
pydantic-settings = "^2.12.0"
dishka = {extras = ["fastapi"], version = "^1.7.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.45"}
asyncmy = "^0.2.11"
cryptography = "^46.0.5"
aioboto3 = "^15.5.0"
filetype = "^1.2.0"
telethon = "^1.42.0"
python-socks = {extras = ["asyncio"], version = "^2.8.0"}
aio-pika = "^9.5.8"
cryptg = "^0.5.2"
structlog = "^25.5.0"
opentelemetry-api = "^1.39.1"
opentelemetry-sdk = "^1.39.1"
opentelemetry-exporter-otlp = "^1.39.1"
opentelemetry-instrumentation-sqlalchemy = "^0.60b1"
opentelemetry-instrumentation-aio-pika = "^0.60b1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-instrumentation-asgi = "^0.60b1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
aiosqlite = {version = ">=0.22.1,<0.23.0"}
types-aioboto3 = {extras = ["essential"], version = "^15.5.0"}
tgcrypto = "^1.2.5"
pyqt5 = "^5.15.11"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[tool.poetry.group.migrations.dependencies]
alembic = {version = ">=1.18.1,<2.0.0"}

[tool.poetry.scripts]
alembic = "parser.migrator:run"
test = "development.test:test"
api = "parser.api:run"
worker = "parser.worker:run"
convert_to_telethon = "development.convert_to_telethon:run"
call_parser = "development.call_parser:run"